generateListForUl done
userListForDrivingLicense() changed to generateListForUl().
With this new name the method get parameters:
first param: the array which we read the values.
second param: the name of the list where we want to display the names.

Need to think about third param to add which property has to be displayed in the list.